Session 2
During this session I fixed clearance issue between the robot plates and made the other side of the robot. Also, I managed to add one of the walls that will have to hold the third odometry wheel but I will still need to add the second wall. Those walls also will be responsible for holding the battery as it is important to mount it close to the ground as it is heavy and it is good to have low center of mass.