Session 3
During this session I created the remaining mount for the third odometry pod and added the pod itself. Then I added the pulleys that I missed for the other side of the drivetrain. And lastly I designed a mount for the control hubs and the battery that securely mounts to the whole chassis with 6 screws and makes the whole thing way more robust while keeping the center of mass close to the ground.