Session 4
During this session I added all of the screws to know if there is enough clearance. Also I Flipped the mount for the control hubs and battery for easier access and when they are facing upwards, there is lower chance of wires scraping the ground. Also I changed out 2 plywood plates in the back of the drivetrain for a U-Channel from GOBILDA. This should make the whole structure way more stiff and less likely to break.